[aiohttp] add TraceTestCase for aiohttp tracing
